tickets/src/routes/__test__/show-ticket.test.ts
import request from 'supertest';
import { app } from '../../app';
import mongoose from 'mongoose';

it('return 200 if all tickets are showing', async () => {
  const title = 'amazing me';
  const price = 12;
  const response = await request(app)
    .post('/api/tickets')
    .set('Cookie', global.signin())
    .send({
      title,
      price,
    })
    .expect(201);
  const ticketResponse = await request(app)
    .get(`/api/tickets/${response.body.id}`)
    .send()
    .expect(200);

  expect(ticketResponse.body.title).toEqual(title);
  expect(ticketResponse.body.price).toEqual(price);
});

it('return 404 if  ticket id does not exists', async () => {
  const id = new mongoose.Types.ObjectId().toHexString();
  await request(app).get(`/api/tickets/${id}`).send({}).expect(404);
});
